requirements.txt
pillow==6.2
absl-py
scikit-image
sklearn
requests
pyyaml
sacred
easydict
opencv-python
davisinteractive
imgaug